+++ plugins/admission/pkg/admission/handler.go
@@ -10,9 +10,9 @@ import (
 	"github.com/sirupsen/logrus"
 	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
 
-	"github.com/fairwindsops/insights-plugins/admission/pkg/models"
-	"github.com/fairwindsops/insights-plugins/admission/pkg/opa"
-	"github.com/fairwindsops/insights-plugins/admission/pkg/polaris"
+	"github.com/fairwindsops/insights-plugins/plugins/admission/pkg/models"
+	"github.com/fairwindsops/insights-plugins/plugins/admission/pkg/opa"
+	"github.com/fairwindsops/insights-plugins/plugins/admission/pkg/polaris"
 )
 
 // Validator is the entry point for the admission webhook.